refactor(routes): consolidate clientLoader logic and remove unused files
