--- src/App/frontend/src/routes/index/stateless-index.loader.ts
@@ -1,21 +1,0 @@
-import { redirect } from 'react-router';
-import type { LoaderFunctionArgs } from 'react-router';
-
-import { GlobalData } from 'src/GlobalData';
-import { isStateless } from 'src/routes/index/isStateless';
-import { getRawFirstPage } from 'src/utils/computeStartUrl';
-
-export function clientLoader({ request }: LoaderFunctionArgs) {
-  if (!isStateless()) {
-    return null;
-  }
-
-  const folderId = GlobalData.applicationMetadata.onEntry?.show;
-  const firstPage = getRawFirstPage(folderId);
-  if (!firstPage) {
-    throw new Error(`Cannot determine start page for stateless app (folderId=${folderId ?? 'undefined'})`);
-  }
-
-  const queryKeys = new URL(request.url).search;
-  return redirect(`/${firstPage}${queryKeys}`);
-}
