+++ src/App/frontend/src/routes/routeModule.ts
@@ -18,6 +18,7 @@ type ConvertedRouteModule = Omit<NonIndexRouteObject, 'path' | 'index' | 'childr
 /**
  * Maps a route module onto the route object shape `createBrowserRouter` expects. React Router's Vite
  * plugin reads the route module exports directly, so this conversion goes away if we decide to adopt it sometime in the future.
+ * Most notably it converts the `default` export to `Component`, and `clientLoader`/`clientAction` to `loader`/`action`.
  *
  * @see https://reactrouter.com/upgrading/router-provider
  */
